Clean up.

+++ engine/src/core/ecs/component/animated_sprite_component.c
@@ -107,7 +107,7 @@ AnimatedSpriteComponent* animated_sprite_component_data_copy_to_animated_sprite(
         animation.isValid = true;
         animation.currentFrame = 0;
         animation.frameCount = animationData->frameCount;
-        for (size_t frameIndex = 0; frameIndex < animationData->frameCount; frameIndex++) {
+        for (size_t frameIndex = 0; (int) frameIndex < animationData->frameCount; frameIndex++) {
             AnimationFrameData* animationFrameData = &animationData->animationFrames[frameIndex];
             AnimationFrame animationFrame;
             animationFrame.texture = rbe_asset_manager_get_texture(animationFrameData->texturePath);
